porto-drf-next/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2"
python-decouple = "^3.6"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
Pillow = "^9.1.0"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
